+++ errortools/partial.py
@@ -2,4 +2,4 @@
 
 from _errortools.partial import __all__ as _all
 
-__all__ = _all # type: ignore
+__all__ = _all  # type: ignore